drm: omapdrm: improve choice of memory type for tiled memory
On omap5/dra7 tiled memory is currently strongly-ordered, which is just
painfully and needlessly slow.  Using device memory suffices and is much
faster.  Even normal uncacheable memory is safe if the application
avoids reading from the framebuffer, so I've added flags to allow
userspace to indicate it is okay with restricted mmap access.

Signed-off-by: Matthijs van Duin <matthijsvanduin@gmail.com>